% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_comments.R
\name{mod_comments_ui}
\alias{mod_comments_ui}
\alias{mod_comments_server}
\title{mod_comments_ui and mod_comments_server}
\usage{
mod_comments_ui(id, episode_df)

mod_comments_server(input, output, session, episode_df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
