% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_episode_box.R
\name{mod_episode_box_ui}
\alias{mod_episode_box_ui}
\alias{mod_episode_box_server}
\title{mod_episode_box_ui and mod_episode_box_server}
\usage{
mod_episode_box_ui(id, episode_title, episode_int, episode_df)

mod_episode_box_server(input, output, session, episode_int, episode_df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
